chore: Extract test helper for docker plugin
Remove some test duplication by extracting a test helper that stubs
the docker plugin response in all docker test. This could benefit other
tests/plugins too, but.. one step at the time.

2 tests were wrapping result in a promise and this appears redundant.
But if it would be needed, it could be easily done in this test helper
consistently for all test cases.