fix: Remove vendored tree-kill dep
We vendored the `tree-kill` dep in `snyk-sbt-plugin` so that we could
temporarily patch it to avoid a vulnerability. Now that it has been
fixed upstream, the change in `snyk-sbt-plugin` has been reverted, and a
new version released with the fixed upstream `tree-kill`.

https://github.com/snyk/snyk-sbt-plugin/pull/73

This patch updates the Snyk CLI with this new release of
`snyk-sbt-plugin`.