fix: Add error code to error object for json/sarif test output
When the --json or --sarif options are used we do not add the error
code in the thrown error, as opposed to when not using these output types.
This results in the error code not being shown in analytics and in logs.
This commit fixes this behavior by adding the error code of the first
found error (as done for non-json/sarif outputs).