test: show that excluding base image vulns returns wrong exit code
The Snyk CLI first produces a not-ok test result when it gets the list of vulnerabilities from Snyk and only then excludes the base image vulns when using the --exclude-base-image-vulns CLI flag. This leads it to incorrectly reporting an error. The CLI sees an unsuccessful test response with 0 vulnerabilities and cannot handle it.