refactor: Strip json result array after formatting
This refactor is done to not have a special case for
array vs single result when formatting the results
for JSON output.
Note: dataToSend will now take the jsonData after formatting
rather than the data before formatting. Since the formatting
only changes the vulnerbilities and removes the unwanted
scanResult, and since the vulnerbilities are removed from
dataToSend before it is used, this change will not affect the
CLI bahvior except for removing the unwanted scanResult.