test: fix test failing on node@10 warning messages
Node 10 support is being deprecated so API alerts have been added server-side to show a warning message in CLI. This isn't relevant to these tests so I've updated the assertions.

We should mock the response but this is blocking pipelines and it would require re-writing the entire test. There are other tests with the same issue so it'd be easier to migrate them together.