feat: container json response with app vulns
We want to support scanning app vulns by default. Before we can do that,
we have to support JSON response without the `experimental` flag, and
without creating any breaking changes for existing users. The solution
is to add another field `applications` to the response object, which
will container all of the app vulns results.