fix: update snyk-docker-plugin
This commit fixes the project names & names for root nodes in the
dependency graph for Go binaries that are built and released by the Go
distribution (e.g. `pprof` or `addr2line`).

It also caused a bug for exactly these binaries from the Go 1.18 release
(Go 1.19 is not affected anymore) which causes scans to fail because the
root node in the depgraph will not contain *any* name at all. This
shouldn't happen and is fixed by bumping the snyk-docker-plugin to
include the last PR.